Get rackup working
This was frustrating. For some reason the rackup command is giving me
all sorts of problems. Explicitly setting the ruby version to 3.2.0
seems to have solved it for now.

This also adds a Gemfile and Gemfile.lock file, and keeps things super
simple according to a basic Roda guide I found:
https://fiachetti.gitlab.io/mastering-roda/